refactor(config): split config into focused modules
Break the monolithic config module into focused UI, places, layout, keys, and loading modules.

Move config tests into dedicated module files to keep coverage alongside the new structure.
